backend/requirements.txt
# backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
pydantic-settings==2.2.1 # <---- ADD THIS LINE (or latest version)
transformers==4.38.2
torch==2.2.1 #--index-url https://download.pytorch.org/whl/cpu
accelerate